requirements.txt

streamlit
openai
openai-whisper
pandas
